scripts/scraper_requirements.txt
# Brand Scraper Requirements
# Install with: pip install -r scraper_requirements.txt

requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
instaloader>=4.10.0
